+++ src/Storages/MergeTree/MergeTreeData.cpp
@@ -56,6 +56,7 @@
 #include <Storages/StorageMergeTree.h>
 #include <Storages/StorageReplicatedMergeTree.h>
 #include <Storages/VirtualColumnUtils.h>
+#include <Storages/Freeze.h>
 #include <Common/Increment.h>
 #include <Common/SimpleIncrement.h>
 #include <Common/Stopwatch.h>
@@ -5957,95 +5958,6 @@ bool MergeTreeData::removeDetachedPart(DiskPtr disk, const String & path, const 
     return false;
 }
 
-// copy and some edits from StorageReplicatedMergeTree::removeSharedDetachedPart
-bool removeSharedDetachedPart(DiskPtr disk, const String & path, const String & part_name, const String & table_uuid,
-    const String &, const String & detached_replica_name, const String & detached_zookeeper_path, Poco::Logger * log, ContextPtr local_context)
-{
-    bool keep_shared = false;
-
-    zkutil::ZooKeeperPtr zookeeper = local_context->getZooKeeper();
-
-    fs::path checksums = fs::path(path) / IMergeTreeDataPart::FILE_FOR_REFERENCES_CHECK;
-    if (disk->exists(checksums))
-    {
-        if (disk->getRefCount(checksums) == 0)
-        {
-            String id = disk->getUniqueId(checksums);
-            keep_shared = !StorageReplicatedMergeTree::unlockSharedDataByID(id, table_uuid, part_name,
-                detached_replica_name, disk, zookeeper, local_context->getReplicatedMergeTreeSettings(), log,
-                detached_zookeeper_path);
-        }
-        else
-            keep_shared = true;
-    }
-
-    disk->removeSharedRecursive(path, keep_shared);
-
-    return keep_shared;
-}
-
-// copy and some edits from StorageReplicatedMergeTree::removeDetachedPart
-bool removeDetachedPart(DiskPtr disk, const String & path, const String &part_name, bool, Poco::Logger * log, ContextPtr local_context)
-{
-    if (disk->supportZeroCopyReplication())
-    {
-        FreezeMetaData meta;
-        if (meta.load(disk, path))
-        {
-            if (meta.is_replicated) {
-                FreezeMetaData::clean(disk, path);
-                return removeSharedDetachedPart(disk, path, part_name, meta.table_shared_id, meta.zookeeper_name, meta.replica_name, "", log, local_context);
-            }
-        }
-    }
-
-    disk->removeRecursive(path);
-
-    return false;
-}
-
-PartitionCommandsResultInfo MergeTreeData::unfreezePartitionsFromTableDirectory(MergeTreeData::MatcherFn matcher, const String & backup_name, Disks disks, fs::path table_directory, Poco::Logger * log, ContextPtr local_context) {
-    PartitionCommandsResultInfo result;
-
-    for (const auto & disk : disks)
-    {
-        if (!disk->exists(table_directory))
-            continue;
-
-        for (auto it = disk->iterateDirectory(table_directory); it->isValid(); it->next())
-        {
-            const auto & partition_directory = it->name();
-
-            /// Partition ID is prefix of part directory name: <partition id>_<rest of part directory name>
-            auto found = partition_directory.find('_');
-            if (found == std::string::npos)
-                continue;
-            auto partition_id = partition_directory.substr(0, found);
-
-            if (!matcher(partition_id))
-                continue;
-
-            const auto & path = it->path();
-
-            bool keep_shared = DB::removeDetachedPart(disk, path, partition_directory, true, log, local_context);
-
-            result.push_back(PartitionCommandResultInfo{
-                .partition_id = partition_id,
-                .part_name = partition_directory,
-                .backup_path = disk->getPath() + table_directory.generic_string(),
-                .part_backup_path = disk->getPath() + path,
-                .backup_name = backup_name,
-            });
-
-            LOG_DEBUG(log, "Unfreezed part by path {}, keep shared data: {}", disk->getPath() + path, keep_shared);
-        }
-    }
-
-    LOG_DEBUG(log, "Unfreezed {} parts", result.size());
-
-    return result;
-}
-
 PartitionCommandsResultInfo MergeTreeData::unfreezePartitionsByMatcher(MatcherFn matcher, const String & backup_name, ContextPtr local_context)
 {
     auto backup_path = fs::path("shadow") / escapeForFileName(backup_name) / relative_data_path;
@@ -6054,7 +5966,7 @@ PartitionCommandsResultInfo MergeTreeData::unfreezePartitionsByMatcher(MatcherFn
 
     auto disks = getStoragePolicy()->getDisks();
 
-    return unfreezePartitionsFromTableDirectory(matcher, backup_name, disks, backup_path, log, local_context);
+    return Unfreezer().unfreezePartitionsFromTableDirectory(matcher, backup_name, disks, backup_path, local_context);
 }
 
 bool MergeTreeData::canReplacePartition(const DataPartPtr & src_part) const
